packages/promptlint/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptlint"
version = "0.3.0"
description = "Static analysis tool for LLM prompts — treats prompts as code"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Scott Converse"},
]
keywords = ["llm", "prompt", "linter", "static-analysis", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "prompttools-core>=1.0,<2.0",
    "typer[all]>=0.12",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tiktoken>=0.7",
    "rich>=13.0",
    "watchfiles>=0.21",
]

[project.scripts]
promptlint = "promptlint.cli:app"

[project.urls]
Homepage = "https://github.com/scottconverse/promptlint"
Repository = "https://github.com/scottconverse/promptlint"
Issues = "https://github.com/scottconverse/promptlint/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src/promptlint"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
